refactor database workers, in order to minimize code repetition.
