+++ src/main/scala/org/gammf/collabora/database/actors/worker/DBWorkerAuthenticationActor.scala
@@ -22,16 +22,7 @@ class DBWorkerAuthenticationActor(override val yellowPages: ActorRef, override v
                                   override val topic: ActorTopic, override val service: ActorService)
   extends UsersDBWorker[DBWorkerMessage] with DefaultDBWorker with Stash {
 
-  override def receive: Receive = ({
-    //TODO consider: these three methods in super class?
-    case message: RegistrationResponseMessage => getActorOrElse(Topic() :+ Database, ConnectionHandler, message)
-      .foreach(_ ! AskConnectionMessage())
-
-    case m: GetConnectionMessage =>
-      connection = Some(m.connection)
-      unstashAll()
-
-    case _ if connection.isEmpty => stash()
+  override def receive: Receive = super.receive orElse ({
 
     case message: LoginMessage =>
       find(
@@ -51,5 +42,5 @@ class DBWorkerAuthenticationActor(override val yellowPages: ActorRef, override v
       ) pipeTo sender
 
     case _ => unhandled(_)
-  }: Receive) orElse super[UsersDBWorker].receive
+  }: Receive)
 }
