+++ src/main/scala/org/gammf/collabora/database/actors/worker/DBWorkerCheckMemberExistenceActor.scala
@@ -17,16 +17,7 @@ class DBWorkerCheckMemberExistenceActor(override val yellowPages: ActorRef, over
                                         override val topic: ActorTopic, override val service: ActorService)
   extends UsersDBWorker[DBWorkerMessage] with DefaultDBWorker with Stash {
 
-  override def receive: Receive = ({
-    //TODO consider: these three methods in super class?
-    case message: RegistrationResponseMessage => getActorOrElse(Topic() :+ Database, ConnectionHandler, message)
-      .foreach(_ ! AskConnectionMessage())
-
-    case message: GetConnectionMessage =>
-      connection = Some(message.connection)
-      unstashAll()
-
-    case _ if connection.isEmpty => stash()
+  override def receive: Receive = super.receive orElse ({
 
     case message: IsMemberExistsMessage =>
       find(
@@ -34,5 +25,5 @@ class DBWorkerCheckMemberExistenceActor(override val yellowPages: ActorRef, over
         okStrategy = bsonDocumet => QueryOkMessage(IsMemberExistsResponseMessage(message.username, bsonDocumet.isDefined)),
         failStrategy = defaultDBWorkerFailStrategy(message.username)
       ) pipeTo sender
-  }: Receive) orElse super[UsersDBWorker].receive
+  }: Receive)
 }
