+++ src/main/scala/org/gammf/collabora/database/actors/worker/DBWorkerCollaborationsActor.scala
@@ -1,18 +1,14 @@
 package org.gammf.collabora.database.actors.worker
 
-import akka.actor.{ActorRef, Stash}
+import akka.actor.ActorRef
 import akka.pattern.pipe
 import org.gammf.collabora.database._
 import org.gammf.collabora.database.messages._
 import org.gammf.collabora.util.Collaboration
-import reactivemongo.api.Cursor
 import reactivemongo.bson.{BSON, BSONDocument, BSONObjectID}
 
 import scala.concurrent.ExecutionContext.Implicits.global
-import org.gammf.collabora.yellowpages.util.Topic
-import org.gammf.collabora.yellowpages.TopicElement._
 import org.gammf.collabora.yellowpages.ActorService._
-import org.gammf.collabora.yellowpages.messages.RegistrationResponseMessage
 import org.gammf.collabora.yellowpages.util.Topic.ActorTopic
 
 /**
@@ -20,18 +16,9 @@ import org.gammf.collabora.yellowpages.util.Topic.ActorTopic
   */
 class DBWorkerCollaborationsActor(override val yellowPages: ActorRef, override val name: String,
                                   override val topic: ActorTopic, override val service: ActorService)
-  extends CollaborationsDBWorker[DBWorkerMessage] with DefaultDBWorker with Stash {
+  extends CollaborationsDBWorker[DBWorkerMessage] with DefaultDBWorker {
 
-  override def receive: Receive = ({
-    //TODO consider: these three methods in super class?
-    case message: RegistrationResponseMessage => getActorOrElse(Topic() :+ Database, ConnectionHandler, message)
-      .foreach(_ ! AskConnectionMessage())
-
-    case message: GetConnectionMessage =>
-      connection = Some(message.connection)
-      unstashAll()
-
-    case _ if connection.isEmpty => stash()
+  override def receive: Receive =  super.receive orElse ({
 
     case message: InsertCollaborationMessage =>
       val bsonCollaboration: BSONDocument = BSON.write(message.collaboration) // necessary conversion, sets the collaborationID
@@ -56,14 +43,7 @@ class DBWorkerCollaborationsActor(override val yellowPages: ActorRef, override v
         failStrategy = defaultDBWorkerFailStrategy(message.userID)
       ) pipeTo sender
 
-      //TODO is this necessary here?
-    case message: GetAllCollaborationsMessage =>
-      getCollaborationsCollection.map(collaborations =>
-        collaborations.find(BSONDocument(
-          COLLABORATION_USERS + "." + COLLABORATION_USER_USERNAME -> message.username
-        )).cursor[BSONDocument]().collect[List](-1, Cursor.FailOnError[List[BSONDocument]]())) // -1 is no limit list
-        .flatten.map(list =>
-        AllCollaborationsMessage(list.map(bson => bson.as[Collaboration]))) pipeTo sender
+    case _ => println("["+ name + "] Huh?"); unhandled(_)
 
-  }: Receive) orElse super[CollaborationsDBWorker].receive
+  }: Receive)
 }
