+++ src/main/scala/org/gammf/collabora/database/actors/worker/DBWorkerMemberActor.scala
@@ -20,16 +20,7 @@ class DBWorkerMemberActor(override val yellowPages: ActorRef, override val name:
                           override val topic: ActorTopic, override val service: ActorService)
   extends CollaborationsDBWorker[DBWorkerMessage] with DefaultDBWorker with Stash {
 
-  override def receive: Receive = ({
-    //TODO consider: these three methods in super class?
-    case message: RegistrationResponseMessage => getActorOrElse(Topic() :+ Database, ConnectionHandler, message)
-      .foreach(_ ! AskConnectionMessage())
-
-    case message: GetConnectionMessage =>
-      connection = Some(message.connection)
-      unstashAll()
-
-    case _ if connection.isEmpty => stash()
+  override def receive: Receive = super.receive orElse ({
 
     case message: InsertMemberMessage =>
       update(
@@ -58,5 +49,5 @@ class DBWorkerMemberActor(override val yellowPages: ActorRef, override val name:
         failStrategy = defaultDBWorkerFailStrategy(message.userID)
       ) pipeTo sender
 
-  }: Receive) orElse super[CollaborationsDBWorker].receive
+  }: Receive)
 }
